Fix remote sh output

+++ Jenkinsfile
@@ -10,19 +10,6 @@ pipeline {
                 git branch: 'main', credentialsId: 'jenkins-github-key', url: 'git@github.com:disasstor/ytdl-flask.git'
 			}
 		}
-		stage('Clearing after PreBuild') {
-            steps {
-				script {
-					def build_images_id_list = (sh(returnStdout: true, script: "docker images | grep ilyatrof/ytdl-flask | awk '{ print \$3 }'")).replace("\n", " ")
-					if (build_images_id_list) {
-						sh """
-						echo Clearing after PreBuild
-						docker rmi -f $build_images_id_list
-						"""
-					}
-				}
-			}
-		}
         stage('Build') {
             steps {
                 sh """
@@ -33,33 +20,33 @@ pipeline {
 				"""
             }
         }
-        stage('Deploy') {
+        stage('PreDeploy') {
             steps {
 				script {
-                    try{
-						def old_container_id_list = (sh(returnStdout: true, script: "docker --host $DOCKER_HOST ps -a | grep ilyatrof/ytdl-flask | awk '{ print \$1 }'")).replace("\n", " ")
+					def old_container_id_list = (sh(returnStdout: true, script: "docker --host $DOCKER_HOST ps -a | grep ilyatrof/ytdl-flask | awk '{ print \$1 }'")).replace("\n", " ")
+					if (old_container_id_list) {	
 						sh """
 						echo Try to kill cootainers
 						docker --host $DOCKER_HOST kill $old_container_id_list
 						"""
-                    }catch (err) {
-                        sh 'echo Kill older cootainers ERROR'
                     }
-                    try{
-						def old_images_id_list = (sh(returnStdout: true, script: "docker --host $DOCKER_HOST images | grep ilyatrof/ytdl-flask | awk '{ print \$3 }'")).replace("\n", " ")
+					def old_images_id_list = (sh(returnStdout: true, script: "docker --host $DOCKER_HOST images | grep ilyatrof/ytdl-flask | awk '{ print \$3 }'")).replace("\n", " ")
+					if (old_images_id_list) {	
 						sh """
 						echo Try to remove image
 						docker --host $DOCKER_HOST rmi -f $old_images_id_list
 						"""
-					}catch (err) {
-                        sh 'echo Remove older image ERROR'
-                    }
+					}
                 }
-                sh """
+            }
+        }
+		stage('Deploy') {
+			steps {
+				sh """
 				echo Deploy new container
                 docker --host $DOCKER_HOST run --rm --name ytdl-flask-app -d -p 5000:5000 ilyatrof/ytdl-flask:v${BUILD_NUMBER}
 				"""
-            }
-        }
+			}
+		}
 	}
 }
